+++ src/meico/mei/Helper.java
@@ -276,6 +276,44 @@ public class Helper {
     }
 
     /**
+     * check if all provided nodes have the same parent
+     * @param nodes
+     * @return
+     */
+    public static boolean haveSameParent(Nodes nodes) {
+        if (nodes == null)
+            return false;
+        if (nodes.size() < 2)
+            return true;
+
+        ParentNode parent = nodes.get(0).getParent();
+        for (int i = 1; i < nodes.size(); ++i) {
+            if (parent != nodes.get(i).getParent())
+                return false;
+        }
+        return true;
+    }
+
+    /**
+     * make a comparison of two elements; ignore their child elements
+     * @param a
+     * @param b
+     * @return true if equal, otherwise false
+     */
+    public static boolean compareFlatElements(Element a, Element b) {
+        if ((a == null) || (b == null)) return false;
+        if (!a.getLocalName().equals(b.getLocalName())) return false;
+
+        for (int i=0; i < a.getAttributeCount(); ++i) {
+            Attribute aAtt = a.getAttribute(i);
+            Attribute bAtt = b.getAttribute(aAtt.getLocalName());
+            if (bAtt == null) return false;
+            if (!aAtt.getValue().equals(bAtt.getValue())) return false;
+        }
+        return true;
+    }
+
+    /**
      * this method adds element addThis to a timely sequenced list, the map, and ensures the timely order of the elements in the map;
      * therefore, addThis must contain the attribute "date"; if not, addThis is appended at the end
      * @param addThis an xml element (should have an attribute date)