+++ src/meico/mpm/elements/styles/ArticulationStyle.java
@@ -6,6 +6,8 @@ import nu.xom.Element;
 
 import java.util.HashMap;
 import java.util.LinkedList;
+import java.util.Map;
+import java.util.UUID;
 
 /**
  * This class interfaces articulation style definitions.